requirements.txt
Flask==2.3.3
requests>=2.31.0
gunicorn>=20.1.0
msal==1.25.0
Flask-Session==0.6.0
cachelib>=0.9.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0